Split live control into focused modules
